refactor: remove any type props
