+++ frontend/src/component/map/SeromFloor.tsx
@@ -1,29 +1,46 @@
 import "./seromFloor.css";
-import { lentInfo, cabinetInfo } from "../../types/cabinetTypes";
+import { cabinetInfo, locationInfo } from "../../redux/slices/cabinetSlice";
+import { lentInfo } from "../../redux/slices/lentSlice";
 
-export function SeromFloor(props: any) {
+interface SeromFloorProps {
+  info: locationInfo;
+  l_idx: number;
+  f_idx: number;
+  floor_name: number;
+  lent: lentInfo[];
+}
+
+export function SeromFloor(props: SeromFloorProps) {
   const cluster1: number = props.f_idx === 1 ? 3 : 5;
   const cluster2: number = props.f_idx === 1 ? 4 : 6;
 
   const findIdx = (name: string) => {
-    const idx = props.info?.section[props.l_idx][props.f_idx].findIndex(
+    // const idx = props.info?.section[props.l_idx][props.f_idx].findIndex(
+    //   (section: string) => section === name
+    // );
+    const idx: number | undefined = props.info?.section?.[props.l_idx][props.f_idx].findIndex(
       (section: string) => section === name
     );
-    if (idx === -1) {
+    if (idx === -1 || idx === undefined) {
+      // FIXME Before (idx === -1)
+      // info의 타입 locationInfo 내부 프로퍼티들이 전부 선택적 프로퍼티임
+      // props.info?.section 에 옵셔널 체이닝으로 section?. 으로 수정
+      // 이로 인해 idx에 들어올 수 있는 값이 number | undefined 로 바뀜
+      // 그래서 조건에 idx === undefined 추가했는데 이게 맞나..요?
       return 1;
     } else {
       return idx + 1;
     }
   };
-  const countCabinet = (name: string): number => {
-    return props.info?.cabinet[props.l_idx][props.f_idx][findIdx(name) - 1]
+  const countCabinet = (name: string): number | undefined => {
+    return props.info?.cabinet?.[props.l_idx][props.f_idx][findIdx(name) - 1]
       .length;
   };
   const countLentCabinet = (name: string): number => {
     let count = 0;
     props.lent?.forEach((cabi: lentInfo) => {
       if (props.info) {
-        props.info.cabinet[props.l_idx][props.f_idx][findIdx(name) - 1].forEach(
+        props.info.cabinet?.[props.l_idx][props.f_idx][findIdx(name) - 1].forEach(
           (cabinet: cabinetInfo) => {
             if (cabinet.cabinet_id === cabi.lent_cabinet_id) {
               count++;
